+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added `-o`/`--out` option to specify output file for each formatter.
 * Added `--require` option to inject external Ruby code into RuboCop.
 * Added `-V/--verbose-version` option that displays Parser version and Ruby version as well.
+* Added `-R/--rails` option that enables extra Rails-specific cops.
 
 ### Changes